+++ src/main/java/com/innowise/skynet/entity/PartType.java
@@ -1,4 +1,4 @@
-package com.innowise.skynet.enums;
+package com.innowise.skynet.entity;
 
 public enum PartType {
     HEAD, TORSO, HAND, FOOT;
